Rename NotesEditor to ChildNotes
